Boss Room appearance is made
You can enter the room and see the maze but the player can still walk all over it.

There needs to be wall objects over the maze wall to stop the player but functionality was behaving strangely.

Also the warp tile from the Boss Room back into room2 does not seem to want to warp back into room2.