Implemented cutscenes for chapter introduction and elevator opening
Commit applies to tasks #125 & #126 and story #112
Have added a proper cutscene to begin the game. Title of initial chapter is 'arrival.' Elevator door animation was implemented and a door zoom in cue was used to give a 3d effect. There are smooth fade-in transitions between these joined cutscenes and starts the game smoothly.