Fixed Game Breaking Bug
There was an issue incurred during merges that led to the development of a game-breaking bug that was passed along the two branches merged prior. It was because of this bug that saving in the Boss Room was impossible and IDs were corrupted. This makes sense, as the differing background processes run by this bug would have led to newly created IDs for differing world states. All merges from here on out should not be nearly as difficult.